% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagerConfirm.R
\name{ManagerConfirm_id}
\alias{ManagerConfirm_id}
\title{查询对账单id}
\usage{
ManagerConfirm_id(dms_token, statementNo)
}
\arguments{
\item{dms_token}{第二个参数}

\item{statementNo}{}
}
\value{
两个数的和
}
\description{
查询对账单id
}
\examples{
ManagerConfirm_id
}
