% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManagerConfirm.R
\name{ManagerConfirm_view}
\alias{ManagerConfirm_view}
\title{采购经理查询对账单}
\usage{
ManagerConfirm_view(dms_token)
}
\arguments{
\item{dms_token}{第二个参数}
}
\value{
两个数的和
}
\description{
采购经理查询对账单
}
\examples{
ManagerConfirm_view
}
