% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PurchaserConfirm.R
\name{PurchaserConfirm_delete}
\alias{PurchaserConfirm_delete}
\title{删除临时表}
\usage{
PurchaserConfirm_delete(dms_token)
}
\arguments{
\item{dms_token}{}
}
\value{
两个数的和
}
\description{
删除临时表
}
\examples{
Task_view
}
