% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PurchaserConfirm.R
\name{PurchaserConfirm_update}
\alias{PurchaserConfirm_update}
\title{更新对账单}
\usage{
PurchaserConfirm_update(dms_token)
}
\arguments{
\item{dms_token}{第二个参数}
}
\value{
两个数的和
}
\description{
更新对账单
}
\examples{
Task_view
}
