% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PurchaserConfirm.R
\name{PurchaserConfirm_view}
\alias{PurchaserConfirm_view}
\title{查询对账单}
\usage{
PurchaserConfirm_view(dms_token, FBILLNO)
}
\arguments{
\item{dms_token}{第二个参数}

\item{FBILLNO}{}
}
\value{
两个数的和
}
\description{
查询对账单
}
\examples{
Task_view
}
